backend/requirements.txt
alembic==1.17.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
cffi==2.0.0
cfgv==3.4.0
click==8.3.0
cryptography==46.0.3
distlib==0.4.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.120.4
google-generativeai==0.8.5
filelock==3.20.0
h11==0.16.0
httpx==0.28.1
identify==2.6.15
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
nodeenv==1.9.1
passlib==1.7.4
platformdirs==4.5.0
pre_commit==4.4.0
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
python-dotenv==1.2.1
PyYAML==6.0.3
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
virtualenv==20.35.4
